refactor: move around on apis
Update rp pico 2 w to use released crates
